Rewrite the "post-link" command
"bin/quranbot" has become the main shell script, and all commands are
written in TypeScript and delegated to Deno. This makes it clearer where
shell scripts start and end.

The logic of "post-link" has been altered to post a link to a surah in
sequential order - one surah after the other.